Bind list of named vectors of variable length into data frame

Based on: https://stackoverflow.com/questions/17308551/

ic_bind_list(x)

Arguments

x
list of named vectors

Examples

ic_bind_list(list(ic_vector(ical_example)))

#>                                       PRODID VERSION  CALSCALE  METHOD
#> 1 -//Google Inc//Google Calendar 70.9054//EN     2.0 GREGORIAN PUBLISH
#>            DTSTART            DTEND          DTSTAMP
#> 1 20180809T160000Z 20180809T163000Z 20180810T094100Z
#>                                     UID          CREATED DESCRIPTION
#> 1 1119ejg4vug5758527atjcrqj3@google.com 20180807T133712Z         \\n
#>      LAST-MODIFIED LOCATION SEQUENCE    STATUS                  SUMMARY TRANSP
#> 1 20180807T133712Z                 0 CONFIRMED ical programming mission OPAQUE

ics_file <- system.file("extdata", "england-and-wales.ics", package = "ical")
ics_raw = readLines(ics_file)
x <- lapply(ic_list(ics_raw), function(x) {
  ic_vector(x)
})
ic_df <- ic_bind_list(x)
head(ic_df)

#>   DTEND;VALUE=DATE DTSTART;VALUE=DATE                 SUMMARY
#> 1         20120103           20120102          New Year’s Day
#> 2         20120407           20120406             Good Friday
#> 3         20120410           20120409           Easter Monday
#> 4         20120508           20120507  Early May bank holiday
#> 5         20120605           20120604     Spring bank holiday
#> 6         20120606           20120605 Queen’s Diamond Jubilee
#>                                         UID SEQUENCE          DTSTAMP
#> 1 ca6af7456b0088abad9a69f9f620f5ac-0@gov.uk        0 20180806T114130Z
#> 2 ca6af7456b0088abad9a69f9f620f5ac-1@gov.uk        0 20180806T114130Z
#> 3 ca6af7456b0088abad9a69f9f620f5ac-2@gov.uk        0 20180806T114130Z
#> 4 ca6af7456b0088abad9a69f9f620f5ac-3@gov.uk        0 20180806T114130Z
#> 5 ca6af7456b0088abad9a69f9f620f5ac-4@gov.uk        0 20180806T114130Z
#> 6 ca6af7456b0088abad9a69f9f620f5ac-5@gov.uk        0 20180806T114130Z

x <- lapply(ical_outlook, function(x) {
  ic_vector(x)
})
ic_bind_list(x)

#>                                                         DESCRIPTION
#> 1                                                //2018.erum.io/\\n
#> 2 Ha aaa\\a\\aYau waaa hava praaaaay haard ay aaw ahaa aha aaa aaka
#>                                                                       UID
#> 1 040000008200E00074C5B7101A82E0080000000006245657188FD301000000000000000
#> 2 040000008200E00074C5B7101A82E00800000000F0CCE6F58CC2D301000000000000000
#>                                                      SUMMARY DTSTART;VALUE=DATE
#> 1 In Budapest for European R Users Meeting (eRum) conference           20180514
#> 2                        aaa aaka shara - Laads aauach avaaa               <NA>
#>   DTEND;VALUE=DATE  CLASS PRIORITY          DTSTAMP      TRANSP    STATUS
#> 1         20180518 PUBLIC        5 20180814T134259Z TRANSPARENT CONFIRMED
#> 2             <NA> PUBLIC        5 20180814T134259Z TRANSPARENT CONFIRMED
#>   SEQUENCE                                           LOCATION
#> 1        0                                                   
#> 2        6 Leeds City Centre - time and venue to be confirmed
#>   X-MICROSOFT-CDO-APPT-SEQUENCE X-MICROSOFT-CDO-BUSYSTATUS
#> 1                             0                       FREE
#> 2                             6                       FREE
#>   X-MICROSOFT-CDO-INTENDEDSTATUS DTSTART;TZID=GMT Standard Time
#> 1                           BUSY                           <NA>
#> 2                           BUSY                20180515T100000
#>   DTEND;TZID=GMT Standard Time
#> 1                         <NA>
#> 2              20180515T160000